Bug: Fix error client help request button throwing white label error
Now re-route the controller to the base order page.

+++ src/main/java/uk/ac/rhul/cs2810/restaurantsystem/model/Notification.java
@@ -26,7 +26,7 @@ public class Notification {
      * The Table requesting assistance.
      */
     @Column(name = "tableNo", nullable = false)
-    private int tableNo;
+    private Long tableNo;
 
     /**
      * A generic message which will be sent to the waiter.
